feature(fix): fix failing tests
