+++ src/utils/helperFunctions.js
@@ -1,3 +1,5 @@
+import ReactHtmlParser from "react-html-parser";
+
 /**
  * @description check if value is empty
  * @param {*} value - input to validate
@@ -21,9 +23,46 @@ export const capitalize = string =>
 
 export const parseURL = (parmName, parm) => parm.split(parmName)[1];
 
+/**
+ *
+ * @param {string} authorName -The username of the author of the article
+ * @param {object} currentUser -A user who is logged in
+ * @returns {boolean} whether current user is the author of the article
+ */
+
 export const isCurrentUserAuthor = (authorName, currentUser) => {
   if (currentUser && authorName === currentUser.username) {
     return true;
   }
   return false;
+};
+
+/**
+ *
+ * @param {string} text - body of the article fetched from database
+ * @returns {object} - the object containing the body of article, and background image of the article
+ */
+
+export const stringToHtmlElement = text => {
+  const body = ReactHtmlParser(text);
+  let firstImage = "https://picsum.photos/200/300/?random";
+  const imageRegEx = /<img .*?>/g;
+
+  const images = text.match(imageRegEx);
+  if (images && images.length) {
+    const oneImage = images[0].match(/src\s*=\s*"(.+?)"/);
+    [firstImage, , ,] = oneImage[0].match(/(["'])(?:(?=(\\?))\2.)*?\1/);
+  }
+  return { body, firstImage };
+};
+
+/**
+ *
+ * @param {string} time - Transform time from ISOString to DateString
+ * @returns {string} - the time to DateString format.
+ */
+
+export const calculateTimeStamp = time => {
+  const date = new Date(time);
+  return date.toDateString();
 };
