--- src/utils/helpers/calculateTimeStamp.js
@@ -1,4 +1,0 @@
-export default time => {
-  const date = new Date(time);
-  return date.toDateString();
-};
